recongize failures during application deployment
When "retain on failure" is enabled, exceptions thrown in the test
class (from tests, but also lifecycle hooks) are treated as a signal
to not undeploy anything. However, that wasn't the case for failures
during application deployment.

With this commit, failures during application deployment are treated
as a failure for the purpose of "retain on failure". This should make
debugging test failures much easier when something fails to deploy.