refactor @ManualApplicationDeployment
A design mistake in `@ManualApplicationDeployment` was found.
This annotation conflates 2 things: intent to deploy the application
manually, and source of application metadata when the `app-metadata`
Quarkus extension is not present. Sometimes we want to deploy
the application manually, yet we don't need to provide app metadata
explicitly, because the application source is in the test suite
and has the `app-metadata` extension.

This commit splits it into 2 different annotations:

- `@ManualApplicationDeployment` only states "I will deploy the app
  manually" and has no attributes;
- `@CustomAppMetadata` has all the app metadata attributes and
  is meant to be used only if `app-metadata` is not present.