prerequisites for updating to Quarkus 1.9
This mainly means moving from `container-image-s2i` to
`container-image-openshift`. The `openshift` extension,
which used to bring `container-image-s2i`, now brings
`container-image-openshift`, so wherever we used to use
`container-image-s2i` explicitly (= the `http` test),
we now use `container-image-openshift` instead.

That mainly means one important change, which is largely
invisible to people who exclusively use Quarkus tooling:
the expectation now is that the application files are
present in `/deployments/target` in the container image.
Previously, they were in `/deployments`.

In addition to that, this commit also deals with one
change in Fabric8 Kubernetes Client: they have split
their `Readiness` class into two (Kubernetes-specific
and OpenShift-specific). As a response, this commit
adds `ReadinessUtil` which unifies that.